+++ src/Network/Mail.php
@@ -110,8 +110,8 @@ class Mail
     
     private function setBoundary()
     {
-        $this->sBoundary = uniqid('Pabana-Mail-') . '-' . md5(rand());
-        $this->sBoundaryAlt = uniqid('Pabana-Mail-') . '-' . md5(rand());
+        $this->sBoundary = uniqid('Pabana') . '-' . md5(rand());
+        $this->sBoundaryAlt = uniqid('Pabana-alt') . '-' . md5(rand());
     }
     
     public function getSender()
@@ -119,7 +119,7 @@ class Mail
         if (!empty($this->armSender)) {
             $sSender = '';
             if (!empty($this->armSender[1])) {
-                $sSender .= '"' . $this->armSender[1] . '" ';
+                $sSender .= $this->addEncodeTag('"' . $this->armSender[1] . '"') . ' ';
             }
             $sSender .= '<' . $this->armSender[0] . '>';
             return $sSender;
@@ -133,7 +133,7 @@ class Mail
         if (!empty($this->armReply)) {
             $sReply = '';
             if (!empty($this->armReply[1])) {
-                $sReply .= '"' . $this->armReply[1] . '" ';
+                $sReply .= $this->addEncodeTag('"' . $this->armReply[1] . '"') . ' ';
             }
             $sReply .= '<' . $this->armReply[0] . '>';
             return $sReply;
@@ -165,7 +165,7 @@ class Mail
             foreach ($this->armRecipient[$sRecipientArray] as $armRecipient) {
                 $sRecipient = '';
                 if (!empty($armRecipient[1])) {
-                    $sRecipient .= '"' . $armRecipient[1] . '" ';
+                    $sRecipient .= $this->addEncodeTag('"' . $this->armRecipient[1] . '"') . ' ';
                 }
                 $sRecipient .= '<' . $armRecipient[0] . '>';
                 $armRecipientList[] = $sRecipient;
@@ -195,7 +195,7 @@ class Mail
             $sHeaderContent .= 'Bcc: ' . $sRecipientBcc . PHP_EOL;
         }
         if (!empty($this->sMailer)) {
-            $sHeaderContent .= 'X-Mailer: ' . $this->_sMailer . PHP_EOL;
+            $sHeaderContent .= 'X-Mailer: ' . $this->sMailer . PHP_EOL;
         }
         $sHeaderContent .= 'MIME-Version: 1.0' . PHP_EOL;
         if (!empty($this->armAttachment)) {
@@ -230,16 +230,21 @@ class Mail
     
     public function send()
     {
-        $sAppEncoding = Configuration::read('application.encoding');
+        $sAppEncoding = strtolower(Configuration::read('application.encoding'));
         $oEncoding = new Encoding();
+        $sSubject = $this->addEncodeTag($this->sSubject);
         $sHeaderContent = $this->getHeaderContent();
-        if ($sAppEncoding != $this->sCharset) {
-            $sHeaderContent = $oEncoding->convert($sHeaderContent, $sAppEncoding, $this->sCharset);
-        }
         $sMailContent = $this->getEmailContent();
         if ($sAppEncoding != $this->sCharset) {
+            $sSubject = $oEncoding->convert($sSubject, $sAppEncoding, $this->sCharset);
+            $sHeaderContent = $oEncoding->convert($sHeaderContent, $sAppEncoding, $this->sCharset);
             $sMailContent = $oEncoding->convert($sMailContent, $sAppEncoding, $this->sCharset);
         }
-        return mail($this->getRecipientTo(), $this->sSubject, $sMailContent, $sHeaderContent);
+        return mail($this->getRecipientTo(), $sSubject, $sMailContent, $sHeaderContent);
+    }
+
+    private function addEncodeTag($sValue)
+    {
+        return '=?' . $this->sCharset . '?Q?' . $sValue . '?=';
     }
 }